Fix avatar pivot in profile
The avatar container's pivot Y was incorrectly set to `AVATAR_BASE_SIZE_DP/2`, causing a slight jump during animation. This commit changes it to `0`. Additionally, the pivot X is now dynamically set to the center of the avatar container during layout updates to ensure correct scaling and positioning. The pivot X and Y for `avatarImage` are commented out as they are not needed.